Reverted

+++ libclc/generic/include/math/clc_ldexp.h
@@ -1,5 +1,3 @@
-#include <clc/clcfunc.h>
-
 _CLC_DEF _CLC_OVERLOAD float __clc_ldexp(float, int);
 
 #ifdef cl_khr_fp64
@@ -9,5 +7,5 @@ _CLC_DEF _CLC_OVERLOAD double __clc_ldexp(double, int);
 
 #ifdef cl_khr_fp16
 #pragma OPENCL EXTENSION cl_khr_fp16 : enable
-_CLC_DEF _CLC_OVERLOAD half __clc_ldexp(half, int);
+_CLC_DEF _CLC_OVERLOAD float __clc_ldexp(half, int);
 #endif
